rot + remove duplicate
