+++ conf/data/registration/fragment3dmatch_dense.yaml
@@ -26,18 +26,12 @@ data:
       params:
         sigma: 0.005
         clip: 0.05
-    - transform: RandomRotate
-      params:
-        degrees: 180
-        axis: 0
-    - transform: RandomRotate
-      params:
-        degrees: 180
-        axis: 1
-    - transform: RandomRotate
-      params:
-        degrees: 180
-        axis: 2
+    - transform: Random3AxisRotation
+        params:
+          apply_rotation: True
+          rot_x: 180
+          rot_y: 180
+          rot_z: 180
     - transform: RandomScaleAnisotropic
       params:
         scales: [0.9,1.2]
@@ -47,15 +41,9 @@ data:
     - transform: FixedPoints
       lparams: [16384]
     - transform: Center
-    - transform: RandomRotate
-      params:
-        degrees: 180
-        axis: 0
-    - transform: RandomRotate
-      params:
-        degrees: 180
-        axis: 1
-    - transform: RandomRotate
-      params:
-        degrees: 180
-        axis: 2
+    - transform: Random3AxisRotation
+        params:
+          apply_rotation: True
+          rot_x: 180
+          rot_y: 180
+          rot_z: 180
