+++ conf/data/registration/fragment3dmatch_partial.yaml
@@ -17,6 +17,7 @@ data:
     - transform: GridSampling
       params:
         size: ${data.first_subsampling}
+        mode: "mean"
   train_transform:
     - transform: SaveOriginalPosId
     - transform: FixedPoints
@@ -25,23 +26,17 @@ data:
       params:
         size: ${data.first_subsampling}
         quantize_coords: False
-        mode: "mean"
+        mode: "last"
     - transform: RandomNoise
       params:
         sigma: 0.005
         clip: 0.05
-    - transform: RandomRotate
-      params:
-        degrees: 180
-        axis: 0
-    - transform: RandomRotate
-      params:
-        degrees: 180
-        axis: 1
-    - transform: RandomRotate
-      params:
-        degrees: 180
-        axis: 2
+    - transform: Random3AxisRotation
+        params:
+          apply_rotation: True
+          rot_x: 180
+          rot_y: 180
+          rot_z: 180
     - transform: RandomScaleAnisotropic
       params:
         scales: [0.9,1.2]
@@ -60,19 +55,13 @@ data:
       params:
         size: ${data.first_subsampling}
         quantize_coords: False
-        mode: "mean"
-    - transform: RandomRotate
-      params:
-        degrees: 180
-        axis: 0
-    - transform: RandomRotate
-      params:
-        degrees: 180
-        axis: 1
-    - transform: RandomRotate
-      params:
-        degrees: 180
-        axis: 2
+        mode: "last"
+    - transform: Random3AxisRotation
+        params:
+          apply_rotation: True
+          rot_x: 180
+          rot_y: 180
+          rot_z: 180
     - transform: AddOnes
     - transform: AddFeatByKey
       params:
